refactor(daemon): move the D-Bus server into facelock-daemon (D6)
The org.facelock.Daemon service — per-method authorization
(Method/Scope/authorize_method), caller-identity resolution, the capture
slot, polkit-gated preview frames, live config reload, idle timeout,
capability drop, and the serve loop — moves from the bin-only facelock-cli
crate to facelock_daemon::server, where integration tests can reach it.

The `facelock daemon` command keeps the process-level concerns: root
check, tracing init, and build_handler (config parse, state layout,
model/EP probes, camera resolution), which leans on CLI-shared modules
(crate::resolved, crate::state_layout). The live reload keeps working
through an injected HandlerRebuild recipe; None disables reload so tests
can construct the service without touching /etc.

Zero wire change: bus name, object path, method signatures, in-band
error encoding, and the dbus/ policy + activation files are untouched
(they reference /usr/bin/facelock, whose CLI surface is unchanged).

facelock-daemon gains zbus (tokio feature), tokio, futures-util, libc,
thiserror; facelock-cli sheds its direct tokio/futures-util deps. The
AGENTS.md dependency table is updated accordingly.
