+++ crates/facelock-cli/Cargo.toml
@@ -36,8 +36,6 @@ libc = "0.2"
 nix = { version = "0.31", features = ["user"] }
 facelock-tpm = { path = "../facelock-tpm" }
 zbus = { version = "5", default-features = false, features = ["tokio", "blocking-api"] }
-tokio = { version = "1", features = ["rt-multi-thread", "macros", "signal"] }
-futures-util = "0.3"
 tss-esapi = { workspace = true, optional = true }
 zeroize = "1"
 bytemuck = "1"
