+++ crates/facelock-daemon/Cargo.toml
@@ -23,6 +23,13 @@ facelock-tpm = { path = "../facelock-tpm" }
 bytemuck = { version = "1", features = ["derive"] }
 serde = { workspace = true }
 serde_json = "1"
+thiserror = { workspace = true }
+# The D-Bus server (src/server.rs) — async service, runtime, and the
+# signal/sleep watchers. libc backs the post-init capability drop.
+zbus = { version = "5", default-features = false, features = ["tokio"] }
+tokio = { version = "1", features = ["rt-multi-thread", "macros", "signal", "time"] }
+futures-util = "0.3"
+libc = "0.2"
 
 [features]
 tpm = ["facelock-tpm/tpm"]
